pyproject.toml
[project]
dependencies = ["sqlalchemy", "aiogram", "pyyaml"]